+++ core-ng/src/main/java/core/framework/impl/web/route/PathNode.java
@@ -1,25 +1,19 @@
 package core.framework.impl.web.route;
 
 import core.framework.api.util.Exceptions;
-import core.framework.api.util.Lists;
 import core.framework.api.util.Maps;
 import core.framework.api.util.Strings;
 import core.framework.impl.web.request.PathParams;
 
-import java.util.List;
 import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * @author neo
  */
 class PathNode {
-    private static final Pattern DYNAMIC_PATH_PATTERN = Pattern.compile("\\:(\\w+)(\\(([^\\(\\)]*)\\)){0,1}");
-
-    private final Map<String, PathNode> staticNodes = Maps.newHashMap();
-    private final List<DynamicNode> dynamicNodes = Lists.newArrayList();
     protected URLHandler handler;
+    private Map<String, PathNode> staticNodes;
+    private DynamicNode dynamicNode;
     private DynamicNode wildcardNode;
 
     URLHandler register(String pathPattern) {
@@ -31,47 +25,41 @@ class PathNode {
             if (handler == null) handler = new URLHandler(pathPattern);
             return handler;
         } else if (currentPath.value.startsWith(":")) {
-            Matcher matcher = DYNAMIC_PATH_PATTERN.matcher(currentPath.value);
-            if (!matcher.matches())
-                throw Exceptions.error("path param must follow :name or :name(regex|*), path={}", pathPattern);
-            String name = matcher.group(1);
-            String pattern = matcher.group(3);
-            return registerDynamicNode(pathPattern, currentPath, name, pattern);
+            int paramIndex = currentPath.value.indexOf('(');
+            int endIndex = paramIndex > 0 ? paramIndex : currentPath.value.length();
+            String name = currentPath.value.substring(1, endIndex);
+            boolean wildcard = paramIndex > 0;
+            if (wildcard) {
+                return registerWildcardNode(pathPattern, currentPath, name);
+            } else {
+                return registerDynamicNode(pathPattern, currentPath, name);
+            }
         } else {
+            if (staticNodes == null) staticNodes = Maps.newHashMap();
             PathNode staticNode = staticNodes.computeIfAbsent(currentPath.value, key -> new PathNode());
             return staticNode.register(pathPattern, currentPath.next);
         }
     }
 
-    private URLHandler registerDynamicNode(String pathPattern, Path currentPath, String name, String pattern) {
-        if ("*".equals(pattern)) {
-            if (currentPath.next != null)
-                throw Exceptions.error("wildcard must at end of path pattern, path={}", pathPattern);
-            if (wildcardNode != null && !Strings.equals(wildcardNode.param, name))
-                throw Exceptions.error("conflict dynamic pattern found, path={}, param={}, conflictedParam={}",
-                    pathPattern, name, wildcardNode.param);
-
-            wildcardNode = new DynamicNode(name, null);
-            return wildcardNode.register(pathPattern, currentPath.next);
+    private URLHandler registerWildcardNode(String pathPattern, Path currentPath, String name) {
+        if (currentPath.next != null) throw Exceptions.error("wildcard must be at end of path pattern, path={}", pathPattern);
+        if (wildcardNode != null) {
+            if (!Strings.equals(wildcardNode.param, name))
+                throw Exceptions.error("found conflict dynamic pattern, path={}, param={}, conflictedParam={}", pathPattern, name, wildcardNode.param);
         } else {
-            DynamicNode dynamicNode = getOrCreateDynamicNode(pathPattern, name, pattern);
-            return dynamicNode.register(pathPattern, currentPath.next);
+            wildcardNode = new DynamicNode(name);
         }
+        return wildcardNode.register(pathPattern, currentPath.next);
     }
 
-    private DynamicNode getOrCreateDynamicNode(String pathPattern, String paramName, String paramPattern) {
-        for (DynamicNode dynamicNode : dynamicNodes) {
-            if ((paramPattern == null && dynamicNode.pattern == null)
-                || (paramPattern != null && paramPattern.equals(dynamicNode.pattern.pattern()))) {
-                if (!Strings.equals(dynamicNode.param, paramName))
-                    throw Exceptions.error("conflict dynamic pattern found, path={}, param={}, conflictedParam={}", pathPattern, paramName, dynamicNode.param);
-                return dynamicNode;
-            }
+    private URLHandler registerDynamicNode(String pathPattern, Path currentPath, String name) {
+        if (dynamicNode != null) {
+            if (!Strings.equals(dynamicNode.param, name))
+                throw Exceptions.error("found conflict dynamic pattern, path={}, param={}, conflictedParam={}", pathPattern, name, dynamicNode.param);
+        } else {
+            dynamicNode = new DynamicNode(name);
         }
-
-        DynamicNode node = new DynamicNode(paramName, paramPattern);
-        dynamicNodes.add(node);
-        return node;
+        return dynamicNode.register(pathPattern, currentPath.next);
     }
 
     URLHandler find(String path, PathParams pathParams) {
@@ -82,14 +70,11 @@ class PathNode {
         Path nextPath = currentPath.next;
         if (nextPath == null) return handler;
 
-        PathNode nextNode = staticNodes.get(nextPath.value);
-        if (nextNode != null) {
-            URLHandler handler = nextNode.find(nextPath, pathParams);
-            if (handler != null) return handler;
-        }
+        URLHandler handler = findStatic(nextPath, pathParams);
+        if (handler != null) return handler;
 
         if (!"/".equals(nextPath.value)) {  // dynamic node should not match trailing slash
-            URLHandler handler = findDynamic(nextPath, pathParams);
+            handler = findDynamic(nextPath, pathParams);
             if (handler != null) return handler;
         }
 
@@ -101,14 +86,22 @@ class PathNode {
         return null;
     }
 
+    private URLHandler findStatic(Path nextPath, PathParams pathParams) {
+        if (staticNodes != null) {
+            PathNode nextNode = staticNodes.get(nextPath.value);
+            if (nextNode != null) {
+                return nextNode.find(nextPath, pathParams);
+            }
+        }
+        return null;
+    }
+
     private URLHandler findDynamic(Path nextPath, PathParams pathParams) {
-        for (DynamicNode dynamicNode : dynamicNodes) {
-            if (dynamicNode.match(nextPath.value)) {
-                URLHandler handler = dynamicNode.find(nextPath, pathParams);
-                if (handler != null) {
-                    pathParams.put(dynamicNode.param, nextPath.value);
-                    return handler;
-                }
+        if (dynamicNode != null) {
+            URLHandler handler = dynamicNode.find(nextPath, pathParams);
+            if (handler != null) {
+                pathParams.put(dynamicNode.param, nextPath.value);
+                return handler;
             }
         }
         return null;
@@ -116,16 +109,9 @@ class PathNode {
 
     static class DynamicNode extends PathNode {
         final String param;
-        final Pattern pattern;
 
-        DynamicNode(String param, String pattern) {
+        DynamicNode(String param) {
             this.param = param;
-            if (pattern == null) this.pattern = null;
-            else this.pattern = Pattern.compile(pattern);
-        }
-
-        boolean match(String path) {
-            return pattern == null || pattern.matcher(path).matches();
         }
     }
 }