+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 4.12.2 (4/19/2017 - )
 * web: action log uses context.requestURL instead of context.path
+* route: remove regex path variable support, not used anymore
 
 ### 4.12.1 (4/10/2017 - 4/18/2017)
 * api: rename WebServiceRequestSigner to WebServiceRequestInterceptor